+++ src/Nito.Disposables/ReferenceCountedDisposable.cs
@@ -19,73 +19,4 @@ namespace Nito.Disposables
             where T : class, IDisposable
             => new ReferenceCountedDisposable<T>(new ReferenceCounter<T>(disposable));
     }
-
-    internal sealed class ReferenceCountedDisposable<T> : SingleDisposable<IReferenceCounter<T>>, IReferenceCountedDisposable<T>
-        where T : class, IDisposable
-    {
-        /// <summary>
-        /// Initializes a reference counted disposable that refers to the specified reference count. The specified reference count must have already been incremented for this instance.
-        /// </summary>
-        public ReferenceCountedDisposable(IReferenceCounter<T> referenceCounter)
-            : base(referenceCounter)
-        {
-        }
-
-        protected override void Dispose(IReferenceCounter<T> referenceCounter)
-        {
-            referenceCounter.TryDecrementCount()?.Dispose();
-        }
-
-        T IReferenceCountedDisposable<T>.Target => ReferenceCounter.TryGetTarget() ?? throw new ObjectDisposedException(nameof(ReferenceCountedDisposable<T>));
-
-        IReferenceCountedDisposable<T> IReferenceCountedDisposable<T>.AddReference()
-        {
-            var referenceCounter = ReferenceCounter;
-            if (!referenceCounter.TryIncrementCount())
-                throw new ObjectDisposedException(nameof(ReferenceCountedDisposable<T>));
-            return new ReferenceCountedDisposable<T>(referenceCounter);
-        }
-
-        IWeakReferenceCountedDisposable<T> IReferenceCountedDisposable<T>.AddWeakReference() => new WeakReferenceCountedDisposable<T>(ReferenceCounter);
-
-        private IReferenceCounter<T> ReferenceCounter
-        {
-            get
-            {
-                IReferenceCounter<T> referenceCounter = null!;
-                // Implementation note: IncrementCount always "succeeds" in updating the context since it always returns the same instance.
-                // So, we know that IncrementCount will be called at most once. It may also be called zero times if this instance is disposed.
-                if (!TryUpdateContext(x => referenceCounter = x))
-                    throw new ObjectDisposedException(nameof(ReferenceCountedDisposable<T>));
-                return referenceCounter;
-            }
-        }
-    }
-
-    internal sealed class WeakReferenceCountedDisposable<T> : IWeakReferenceCountedDisposable<T>
-        where T : class, IDisposable
-    {
-        private readonly WeakReference<IReferenceCounter<T>> _weakReference;
-
-        public WeakReferenceCountedDisposable(IReferenceCounter<T> referenceCounter)
-        {
-            _weakReference = new(referenceCounter);
-        }
-
-        IReferenceCountedDisposable<T>? IWeakReferenceCountedDisposable<T>.TryAddReference()
-        {
-            if (!_weakReference.TryGetTarget(out var referenceCounter))
-                return null;
-            if (!referenceCounter.TryIncrementCount())
-                return null;
-            return new ReferenceCountedDisposable<T>(referenceCounter);
-        }
-
-        T? IWeakReferenceCountedDisposable<T>.TryGetTarget()
-        {
-            if (!_weakReference.TryGetTarget(out var referenceCounter))
-                return null;
-            return referenceCounter.TryGetTarget();
-        }
-    }
 }
